Update robots.txt and refactor BlogPanel and Blogsite components for improved structure and styling
